refactored symbol table into module
